extract campaign launch route
